diff code files for treatment page
