+++ pro-dental-care-ui/src/app/patient/records/treatment plan/treatment-plan.component.ts
@@ -1,0 +1,11 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "treatment-plan",
+  templateUrl: "./treatment-plan.component.html",
+  styleUrls: ["./treatment-plan.component.css"],
+})
+export class PatientTreatmentPlan {
+  toDoExpanded = false;
+  treatmentPlanExpanded = false;
+}
